Move tornadoql handlers out into modules #51
