+++ maverick_api/modules/base/tornadoql/dev_graphiql_handler.py
@@ -1,0 +1,9 @@
+import tornado.web
+
+class GraphiQLHandler(tornado.web.RequestHandler):
+    def get(self):
+        if options.development:
+            self.render(os.path.join(self.application.settings.get("static_path"), "graphiql.html"))
+        else:
+            self.set_status(403)
+        self.finish()
